fix: persists file handling should be at PROCESS_POSTUPDATE phase (#274)
This PR fixes the following unexpected behavior due to persist file handling is processed in wrong phase(should be in PROCESS_POSTUPDATE, but in APPLY_UPDATE):

1. all files are processed, but no progress is made and phase is still stuck at APPLY_UPDATE phase,
2. if persist file handling also copying the swapfile, APPLY_UPDATE will be stuck with even longer time.

This PR fixes the above issue by move persist file handling out of create_standby package, and implements it in PROCESS_POSTUPDATE phase.
